Split the native engine by responsibility
The pure-Go engine accumulated lifecycle, objects, rich text, transactions,
patches, history, merge and sync in one five-thousand-line backend file, while
state materialization and rich-text anchor logic shared another large file. The
code was difficult to navigate and made unrelated changes collide.

Split the implementation into focused files inside the internal native package:
engine lifecycle, object operations, rich text, transactions, patches, history,
sync and common helpers, plus separate sequence, rich-text and hydration state.
Document the boundaries and dependency direction. This is a mechanical move
with no behavior changes; parity, conformance and race suites remain green.
