+++ pkg/automerge/internal/native/decode_test.go
@@ -31,6 +31,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
 	"go.probo.inc/probo/pkg/automerge/internal/reference"
 )
